feat(cli): adding args, bundle name, chunk naming
Allow the user to overwrite the default output bundle and chunk
filenaming that this package uses. This respects webpack placeholders
such as [name] and [hash].
Add man text for the cli args.
If args are not provided, we default to using what is used now.
If the user omits '.js' extension from args, we apply it automatically.