feat: rewrite to take advantage of webpackDevServer writeToDisk setting
CRA has gone a long way since I wrote the first cra-build-watch version. Their webpack configuration
changed a lot and often broke this tool in subtle ways. It has proven to be hard to maintain while
being brittle and providing only a subset of the features CRA offers out of the box. Since then
webpack-dev-server provided a new feature called writeToDisk which does what cra-build-watch first
intended to provide.  This rewrite intends on making a clean break from the past and embrace the
future 🚀.

BREAKING CHANGE: Remove most of the flags/features of cra-build-watch. It is not anymore a CLI
replacement for react-scripts but rather something you call one-time after an `npm install`  and
then forget about it. It keeps CRA webpack configuration intact so you can enjoy all of CRA's
wonderful features.